Get rid of 'Result'
The problem with this approach was that since I'm not throwing custom exception, exceptions from tests don't make their way up to the top so that test could fail.
I decided to replace 'Result'. My experiments with custom exception type doesn't look very readable and overhead of new class does not worth it