Introduce 'SwatchCount' enum
Having enum instead of List<Int> makes it easier and less bug-prone to refer to a particular item.
If you have list of possible swatch count items like [2, 4, 6, 8], and want to have a particular item as a value by default, then the only way is to hardcode it: val default = 6.
If something changes in the list, the value by default won't change, thus creating a bug.
With enum, you can use reference to the entry. If this entry is gone, the code won't compile.