Move 'route' property to ':app' module
Considering navigation module as a self-sufficient component, there should not be values that required only because the app uses a particular navigation framework

The 'route: String' property wouldn't be there if ':app' module was not using 'androidx.navigation' framework. This is an indicator that this property belongs to the ':app' module.