Abolish injection of initial state (models)
After trying this approach, I think it is redundant.
First, it doesn't look natural.
Second, it creates more components (although yeah, it simplifies tests a little bit).
And third, most important, it is bug prone. It's easy to mess up initial models you set in test and get a SUT that configured in a way that would not be possible in production.
Better to "drive" a SUT into desired state by executing production-plausible operations on it.