Fix bug
When going from Home screen to Settings screen, but not letting the cross-fade animation of the screen change to finish and clicking "back" to go to the Home screen, then navigation bar appearance is lost.
That was happening because screen change animation doesn't finish, so Home screen never leaves a composition, thus it never enters it again to push a nav bar appearance and override a default one that is set when screen changes.