Make each page of 'Color Center' fill the whole width
When "resume from last searched color on app startup" is enabled and app is launched, then there's a one frame sometimes when the "View color scheme >" button is visible but it's not centered. Next frame, it becomes centered horizontally. It creates unnecessary twitching in UI. Making page take all width by default, regardless of the content inside resolves this issue.