Emit 'null' from 'proceedExecutorFlow' when 'ColorCenterComponents' are cleared
Not really unit-testable.
Creating new 'ProceedExecutor' may happen before it's being read in 'proceed()', and in this case everything will work fine.
However, sometimes the order will be different, and 'proceed()' will read a previous, outdated instance of 'ProceedExecutor' which uses outdated Events/Stores.
In unit tests, the order will always be the same and correct due to test coroutine dispatchers. In the real app, with real dispatchers, order is almost random.