Retain (memoize) not-null data of 'Color Center' after it changes to keep 'Color Center' on screen continuously.
Consider case when there's already color in 'Color Center' and user clicks "Randomize color" button.
1. `HomeViewModel` will end current 'Color Center' session because new color will appear in 'Color Input'. This will cause `HomeData.proceedResult` to become `null`.
2. If "auto proceed with randomized colors" feature is enabled, then newly randomized color will be instantly proceeded with and 'Color Center' will start loading data. This will cause `HomeData.proceedResult` to become `ProceedResult.Success`.

Sometimes both steps happen fast enough so that UI doesn't have time to render `HomeData` after 1st step but before 2nd one. UI renders data that was before step 1 and the next frame is data after step 2. This is visually displayed as 'Color Center' not leaving screen for a single frame.
But sometimes UI has enough time to render data that is between steps 1 and 2. This is visually displayed as:
1. frame with 'Color Center' for initial color
2. frame without 'Color Center' once new random color is set (step 1)
3. frame with 'Color Center' for new random color (step 2).

Having 'Color Center' leaving composition for one frame and re-entering right after produces visual flashing in UI. It also makes more difficult animating between two 'Color Center's.
Thus, we want to batch rapid data changes together. It's not a task for ViewModel, because the problem is in UI.
The solution is to use memoized data and update it with actual after some delay, which is cancelled and restarted if actual data changes before delay times out.