Fix bug: app was breaking if clicked "go to exact color" twice rapidly in 'Color Details'
1. Launch app and enter proceed with some non-exact color
2. click "go to exact color" in 'Color Details' twice rapidly
3. observe that "go to initial color" doesn't work and app overall kinda breaks

This was happening due to specifics of orchestrating events in `HomeViewModel`.
When we send a color to 'Color Input', we anticipate it to be emitted back from `ColorInputColorStore`. However, this flow is a `StateFlow`, so if we set the same color twice consecutively, the flow will emit the color to collectors only first time. Thus, we need a mechanism that determines whether the color we're about to set will be emitted back to collectors before setting it. If not, we don't want to add sent color to the list of sent but not yet processed colors, otherwise it'll stay in the list forever.