Use manually created `HomeUiState` instead of the one assembled from collected components
Consider following scenario:
1. Data transaction starts in ViewModel, and `flowOfIsDataBeingUpdated` emits `true`
2. `ColorPreviewViewModel.dataFlow` emits new value
3. Data transaction finishes, and `flowOfIsDataBeingUpdated` emits `false`

Sometimes it may happen that the Flow from step 2 is collected AFTER the Flow from step 3.
This leads to `stabilize()` emitting a `HomeUiState` composed of stale values (because the Flow from step 2 hasn't been collected with the latest emitted value yet).
The solution is to not consider any data from the component Flows and instead create `HomeUiState` manually from actual values of the component flows.