Use `contentKey` provided by default
Values of each `DataState` are *immutable (unlike 'Color Scheme'). Thus, even if `DataState` type stays the same but some of its data changes, then it is a `DataState` for a different content, which requires a transition (which is achieved with a default `contentKey`).

*immutable:
Both `ColorDetailsData` and `ColorSchemeData` are immutable data classes. However, the latter exposes multiple callbacks that slightly alter (create a copy) values of the given `ColorSchemeData`. In comparison, `ColorDetailsData` doesn't "mutate" in the same way, and each instance of `ColorDetailsData` is a completely new data.