Put `edgeToEdge()` modifier higher to reduce measurement cycles
This way `PagerState.layoutInfo` is populated with correct, final measurements the very first time. In comparison, without this change the `PagerState.layoutInfo` is populated twice, first with intermediate values, and then with final ones.
This change makes so that `calcContentPadding()` produces final correct padding on the first attempt (after initial `null`) as well.