Optimize `TrailingButton()` Composable to skip recomposition of inner `IconButton()` due to changing reference to `ClearTextFeature`
`IconButton()` takes `onClick` parameter, which is a `ClearTextFeature::invoke`.
Due to the fact that on each text update a new feature object is created, Compose treats it as usual changing value, thus recomposing the icon button.
The trick is to use inline lambdas as value of `onClick`, because they are remembered by 'strong skipping mode'. Then, to make this lambda look stable to Compose, it must capture the same object across recompositions. Such constant object is Compose's `State`, so wrapping feature into a remembered state and using it in lambda allows to skip button recomposition.

For Compose's optimization for inline lambdas, see:
https://developer.android.com/develop/ui/compose/performance/stability/strongskipping