Accept `initialText` in the constructor to get rid of `dataUpdatesFlow` nullability
Initially, the `initialText` should have been of type `Text`, not `String`. Sadly, it's not possible right now due to Dagger's limitation on resolving mangling. More described in the Kdoc of `TextFieldViewModel.Factory`.

In the attempts to fix it, I've tried updating versions of Kotlin, Gradle, AGP, and KSP, but to no avail. Nonetheless, it's a good change to bump versions of dependencies, so I am keeping it.