Refactor and rename `ImmediateEventRelay` into `ImmediateRelay`
Enforce collection of the values in a View to happen on the `Dispatchers.Main.immediate` by the implementation.

Only expose "flow for View" from 'Color Input' ViewModels.

Use `receiveAsFlow()` to allow resubscribing.
For example, when returning to the previous screen, when View starts collection on the same `FlowForView` instance again, but with the new collector (different instance)